separated authcontext and authprovider into different files
